separate methodcall and method response xml types
